requirements.txt
# requirements.txt
Flask
google-generativeai